modified:   IoCasas/src/main/java/es/upm/dit/isst/iocasasapi/controller/ReservaController.java 	modified:   IoCasas/src/main/java/es/upm/dit/isst/iocasasapi/repository/reservaRepository.java 	modified:   IoCasas/target/classes/es/upm/dit/isst/iocasasapi/controller/ReservaController.class 	modified:   IoCasas/target/classes/es/upm/dit/isst/iocasasapi/repository/reservaRepository.class

+++ IoCasas/src/main/java/es/upm/dit/isst/iocasasapi/repository/reservaRepository.java
@@ -5,6 +5,6 @@ import es.upm.dit.isst.iocasasapi.model.Reserva;
 
 public interface reservaRepository extends CrudRepository<Reserva, Long> {
 
-    List<Reserva> findByIdInquilino(int idInquilino);
+    List<Reserva> findByIdInquilino(Long idInquilino);
 
 }
